a strech goal I've had in mind for a while. maybe one day this bot will be running in multiple discord servers, working for different schools. these schools will have their own schedule system and require different code to read them.
the idea is that no frontend code will have to deal with this. however, I need to split up the schedule-reading code from the schedule-providing code. the schedule-providing code will stay where it is, but all code that currently reads the CSV that comes from my college will go into a separate component. in the future, other schools will (if that ever happens) have their own schedule-reading component, which may not use CSV files at all, but maybe databases or what have you.

there's a bunch of other stuff that needs to be done for this to work properly, such as preventing users from one school from seeing schedules for another. but this is the first thing that needs to be done.