Automatic restart: increased both the timeout and restart delay from 10 seconds to 20.
Explanation: it appears that the client is still trying to reconnect when the timer reaches 10 seconds (even though that probably would fail anyway, because successful reconnects take less than a second to finish), which seems to prevent StopAsync from returning. this causes the new program instance to fail immediately because another instance is running. I'm hoping that the client will have given up after 40 seconds (which is the total time it now takes before a new instance is started), and the old instance will have closed by then.