3rd attempt at removing the Name field from ScheduleReaders
I'm keeping it in ScheduleService because it's used for log messages. However, since we no longer have to give a name that is equal to the name of a ScheduleRecord property, we can now assign whatever we want.

DISCLAIMER: It is 11:28 pm and I have not tested this. Good night.