added RoosterCommandService.AddLocalizedModule
found a way to localize command names. It involves manually building a module from a Type, one for each supported culture, in nearly the same way as Discord.NET does it, and that involves some level of hacking to bypass the library's internal code. But it works, I just need to make a precondition that restricts each version to specific servers. And that's easy.