fixed startup check on linux; simplified exit signal logic.
apparently, on linux the process name tends to be "dotnet" and I realized this is not a good way to be ensuring only one instance of the program exists on a machine. i've switched to using a named system-wide mutex.

as for exit signaling: there's now a single CancellationTokenSource which signals the process to stop, and it is cancelled through various separate means, which does not involve checking 3 conditions every half-second. this should be a bit easier on system resources, and probably easier to understand.